pyapp/requirements.txt
dependency-injector==4.42.0
fastapi==0.115.0
fastapi-restful==0.6.0
faststream[kafka]==0.5.27
gunicorn==23.0.0
loguru==0.7.2
pydantic==2.9.2
pydantic_settings==2.5.2
pymysql==1.1.1
python-dotenv==1.0.1
PyYAML==6.0.2
sqlalchemy==2.0.35
typing-inspect==0.9.0
uvicorn==0.31.0
